saperated watcher from main
