+++ AstroPhotometry/View/MainWindow.xaml.cs
@@ -21,10 +21,10 @@ namespace AstroPhotometry
         {
             InitializeComponent();
             InitializeFileSystemObjects(); // TODO: filter only folders images and fits
-            
+
             // TODO: find the pyhon folder no metter what (maybe copy the content to bin)
             string base_path = System.IO.Path.GetFullPath("../../../python/");
-             
+
             photo = new PhotoVM();
             DataContext = photo;
             var py_runner = new PythonVM(base_path, @".\tmp\");
@@ -34,74 +34,13 @@ namespace AstroPhotometry
             //py_runner.run("helloworld.py", "qwer");
 
             // wather:
-            var watcher = new FileSystemWatcher(System.IO.Path.GetFullPath("."));
-
-            watcher.NotifyFilter = NotifyFilters.Attributes
-                                 | NotifyFilters.CreationTime
-                                 | NotifyFilters.DirectoryName
-                                 | NotifyFilters.FileName
-                                 | NotifyFilters.LastAccess
-                                 | NotifyFilters.LastWrite
-                                 | NotifyFilters.Security
-                                 | NotifyFilters.Size;
-
-            watcher.Changed += OnChanged;
-            watcher.Created += OnCreated;
-            watcher.Deleted += OnDeleted;
-            watcher.Renamed += OnRenamed;
-            watcher.Error += OnError;
-
-            watcher.Filter = "*.fit*"; // The file to watch -> TODO: do it better
-            watcher.IncludeSubdirectories = true;
-            watcher.EnableRaisingEvents = true;
+            var watcher = new ViewModels.FileWatcherVM(".", "*.fit*");
 
         }
 
-        #region watcher
 
-        private static void OnChanged(object sender, FileSystemEventArgs e)
-        {
-            if (e.ChangeType != WatcherChangeTypes.Changed)
-            {
-                return;
-            }
-            Console.WriteLine($"Changed: {e.FullPath}");
-        }
 
-        private static void OnCreated(object sender, FileSystemEventArgs e)
-        {
-            string value = $"Created: {e.FullPath}";
-            Console.WriteLine(value);
-        }
-
-        private static void OnDeleted(object sender, FileSystemEventArgs e) =>
-            Console.WriteLine($"Deleted: {e.FullPath}");
-
-        private static void OnRenamed(object sender, RenamedEventArgs e)
-        {
-            Console.WriteLine($"Renamed:");
-            Console.WriteLine($"    Old: {e.OldFullPath}");
-            Console.WriteLine($"    New: {e.FullPath}");
-        }
-
-        private static void OnError(object sender, ErrorEventArgs e) =>
-            PrintException(e.GetException());
-
-        private static void PrintException(Exception ex)
-        {
-            if (ex != null)
-            {
-                Console.WriteLine($"Message: {ex.Message}");
-                Console.WriteLine("Stacktrace:");
-                Console.WriteLine(ex.StackTrace);
-                Console.WriteLine();
-                PrintException(ex.InnerException);
-            }
-        }
-    
-    #endregion
-
-    private void Button_Click(object sender, RoutedEventArgs e)
+        private void Button_Click(object sender, RoutedEventArgs e)
         {
             // String filePos = textBox_file.Text.Replace("\\", "/");
